Send number of questions as prop to Question
Instead of sending a navigation object which tells the Question whether
to enable or disable the Previous/Next buttons, we're now sending the
number of questions as a prop, and the Question builds the navigation
object from that.

Also, created a seperate component NavigationButton that uses
withRouter to change the history programmatically (the old way with
wrapping a FlatButton with a Link actually still redirected when the
button was disabled, which was unintended).